pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp==1.0.0",
    "openmeteo-requests==1.2.0",
    "requests-cache==1.1.1",
    "requests==2.31.0",
    "retry-requests==2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.1.0,<3.0.0",
    "xarray>=2023.6.0",
    "aiohttp==3.9.1",
    "asyncio-throttle==1.0.2",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.0",
    "isort==5.13.2",
    "loguru==0.7.2",
    "anemoi-inference>=0.1.0",
    "netcdf4>=1.6.0",
    "docker>=6.0.0",
]
